requirements.txt
pandas==2.0.1
numpy==1.24.1
pyarrow==13.0.0
python-dotenv==1.0.0
maritalk